# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Project orchestration operations
  - name: golden-path
    description: Golden Path configuration operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      description: Returns a paginated list of projects ordered by creation date (newest first).
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project from a prompt
      description: Accepts a natural language prompt, creates a project record, and kicks off async AI spec generation following Golden Path templates.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateProjectResponse"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project status and files
      description: Returns project status, spec, generated file tree, and deploy URL. Poll-friendly.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetails"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}/approve-spec:
    post:
      operationId: approveSpec
      tags: [projects]
      summary: Approve the architectural spec and begin code generation
      description: Transitions project from planned to generating status and kicks off code generation using the approved spec.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Spec approved and generation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApproveSpecResponse"
        "400":
          description: Project not in planned status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}/regenerate-spec:
    post:
      operationId: regenerateSpec
      tags: [projects]
      summary: Regenerate the architectural spec from the original prompt
      description: Discards the current spec and triggers a new AI spec generation. Only valid when project status is planned or failed.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Spec regeneration started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegenerateSpecResponse"
        "400":
          description: Project not in a valid status for regeneration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}/update-spec:
    patch:
      operationId: updateSpec
      tags: [projects]
      summary: Update the architectural spec before approval
      description: Allows editing individual sections of the spec while in planned status.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSpecBody"
      responses:
        "200":
          description: Spec updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetails"
        "400":
          description: Project not in planned status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}/deploy:
    post:
      operationId: deployProject
      tags: [projects]
      summary: Deploy a generated project
      description: Packages generated code and deploys it to a live preview URL.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deployment triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeployProjectResponse"
        "400":
          description: Project not ready for deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /golden-path-configs:
    get:
      operationId: listGoldenPathConfigs
      tags: [golden-path]
      summary: List all Golden Path configurations
      description: Returns all saved Golden Path configurations plus the default rules.
      responses:
        "200":
          description: Config list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoldenPathConfigListResponse"
    post:
      operationId: createGoldenPathConfig
      tags: [golden-path]
      summary: Create a new Golden Path configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoldenPathConfigBody"
      responses:
        "201":
          description: Config created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoldenPathConfigResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /golden-path-configs/active:
    get:
      operationId: getActiveGoldenPathConfig
      tags: [golden-path]
      summary: Get the currently active Golden Path configuration
      description: Returns the active config, or the built-in default if none is active.
      responses:
        "200":
          description: Active config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoldenPathConfigResponse"
  /golden-path-configs/reset-to-default:
    post:
      operationId: resetGoldenPathToDefault
      tags: [golden-path]
      summary: Reset to built-in default configuration
      description: Deactivates all custom configs and reverts to the built-in default.
      responses:
        "200":
          description: Reset to default
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoldenPathConfigResponse"
  /golden-path-configs/{id}:
    put:
      operationId: updateGoldenPathConfig
      tags: [golden-path]
      summary: Update a Golden Path configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoldenPathConfigBody"
      responses:
        "200":
          description: Config updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoldenPathConfigResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Config not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteGoldenPathConfig
      tags: [golden-path]
      summary: Delete a Golden Path configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Config deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
        "404":
          description: Config not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /golden-path-configs/{id}/activate:
    post:
      operationId: activateGoldenPathConfig
      tags: [golden-path]
      summary: Activate a Golden Path configuration
      description: Sets this config as the active one, deactivating any previously active config.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Config activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoldenPathConfigResponse"
        "404":
          description: Config not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CreateProjectBody:
      type: object
      properties:
        prompt:
          type: string
          description: Natural language description of the application to build
      required:
        - prompt
    CreateProjectResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [pending, planning, planned, generating, validating, ready, deployed, failed, failed_checks]
      required:
        - id
        - status
    ProjectFile:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
      required:
        - path
        - content
    GoldenPathCheck:
      type: object
      properties:
        name:
          type: string
        passed:
          type: boolean
        description:
          type: string
        critical:
          type: boolean
      required:
        - name
        - passed
        - description
    ApiEndpointSpec:
      type: object
      properties:
        method:
          type: string
        path:
          type: string
        description:
          type: string
      required:
        - method
        - path
        - description
    DatabaseTableSpec:
      type: object
      properties:
        name:
          type: string
        columns:
          type: array
          items:
            type: string
      required:
        - name
        - columns
    ProjectSpec:
      type: object
      properties:
        overview:
          type: string
        fileStructure:
          type: array
          items:
            type: string
        apiEndpoints:
          type: array
          items:
            $ref: "#/components/schemas/ApiEndpointSpec"
        databaseTables:
          type: array
          items:
            $ref: "#/components/schemas/DatabaseTableSpec"
        middleware:
          type: array
          items:
            type: string
        architecturalDecisions:
          type: array
          items:
            type: string
      required:
        - overview
        - fileStructure
        - apiEndpoints
        - databaseTables
        - middleware
        - architecturalDecisions
    PipelineStageStatus:
      type: object
      properties:
        role:
          type: string
          enum: [architect, backend, frontend, security]
        label:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        startedAt:
          type: string
          nullable: true
        completedAt:
          type: string
          nullable: true
        fileCount:
          type: integer
          nullable: true
        error:
          type: string
          nullable: true
      required:
        - role
        - label
        - status
    PipelineStatus:
      type: object
      properties:
        stages:
          type: array
          items:
            $ref: "#/components/schemas/PipelineStageStatus"
        currentAgent:
          type: string
          nullable: true
      required:
        - stages
    ProjectDetails:
      type: object
      properties:
        id:
          type: string
        prompt:
          type: string
        status:
          type: string
          enum: [pending, planning, planned, generating, validating, ready, deployed, failed, failed_checks]
        spec:
          $ref: "#/components/schemas/ProjectSpec"
        files:
          type: array
          items:
            $ref: "#/components/schemas/ProjectFile"
        goldenPathChecks:
          type: array
          items:
            $ref: "#/components/schemas/GoldenPathCheck"
        pipelineStatus:
          $ref: "#/components/schemas/PipelineStatus"
        deployUrl:
          type: string
          nullable: true
        sandboxId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        error:
          type: string
          nullable: true
      required:
        - id
        - prompt
        - status
        - files
        - goldenPathChecks
        - createdAt
    ApproveSpecResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [generating]
      required:
        - id
        - status
    DeployProjectResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [deployed]
        deployUrl:
          type: string
        sandboxId:
          type: string
          nullable: true
      required:
        - id
        - status
        - deployUrl
    ProjectSummary:
      type: object
      properties:
        id:
          type: string
        prompt:
          type: string
        status:
          type: string
          enum: [pending, planning, planned, generating, validating, ready, deployed, failed, failed_checks]
        fileCount:
          type: integer
        goldenPathScore:
          type: string
          description: "Pass rate like '9/9'"
        deployUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        error:
          type: string
          nullable: true
      required:
        - id
        - prompt
        - status
        - fileCount
        - goldenPathScore
        - createdAt
    ProjectListResponse:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/ProjectSummary"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - projects
        - total
        - limit
        - offset
    RegenerateSpecResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [planning]
      required:
        - id
        - status
    UpdateSpecBody:
      type: object
      properties:
        overview:
          type: string
        fileStructure:
          type: array
          items:
            type: string
        apiEndpoints:
          type: array
          items:
            $ref: "#/components/schemas/ApiEndpointSpec"
        databaseTables:
          type: array
          items:
            $ref: "#/components/schemas/DatabaseTableSpec"
        middleware:
          type: array
          items:
            type: string
        architecturalDecisions:
          type: array
          items:
            type: string
    GoldenPathRuleCheck:
      type: object
      properties:
        type:
          type: string
          enum: [file_pattern, content_match, content_not_match]
        pattern:
          type: string
      required:
        - type
        - pattern
    GoldenPathRule:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        promptInstruction:
          type: string
        critical:
          type: boolean
        check:
          $ref: "#/components/schemas/GoldenPathRuleCheck"
      required:
        - name
        - description
        - promptInstruction
        - check
    TechStack:
      type: object
      properties:
        backend:
          type: string
        frontend:
          type: string
        language:
          type: string
        orm:
          type: string
        validation:
          type: string
      required: [backend, frontend, language, orm, validation]
    FolderStructure:
      type: object
      properties:
        backend:
          type: array
          items:
            type: string
        frontend:
          type: array
          items:
            type: string
        shared:
          type: array
          items:
            type: string
        root:
          type: array
          items:
            type: string
      required: [backend, frontend, shared, root]
    SecurityConfig:
      type: object
      properties:
        requireHelmet:
          type: boolean
        requireCors:
          type: boolean
        requireRateLimiting:
          type: boolean
        noHardcodedSecrets:
          type: boolean
      required: [requireHelmet, requireCors, requireRateLimiting, noHardcodedSecrets]
    CodeQualityConfig:
      type: object
      properties:
        strictTypeScript:
          type: boolean
        noAnyTypes:
          type: boolean
        explicitReturnTypes:
          type: boolean
        esmImports:
          type: boolean
      required: [strictTypeScript, noAnyTypes, explicitReturnTypes, esmImports]
    DatabaseConfig:
      type: object
      properties:
        requireSchema:
          type: boolean
        requireConnectionPooling:
          type: boolean
        requireParameterizedQueries:
          type: boolean
      required: [requireSchema, requireConnectionPooling, requireParameterizedQueries]
    ErrorHandlingConfig:
      type: object
      properties:
        requireGlobalHandler:
          type: boolean
        structuredResponses:
          type: boolean
        noStackTraceLeaks:
          type: boolean
      required: [requireGlobalHandler, structuredResponses, noStackTraceLeaks]
    GoldenPathConfigRules:
      type: object
      properties:
        techStack:
          $ref: "#/components/schemas/TechStack"
        folderStructure:
          $ref: "#/components/schemas/FolderStructure"
        security:
          $ref: "#/components/schemas/SecurityConfig"
        codeQuality:
          $ref: "#/components/schemas/CodeQualityConfig"
        database:
          $ref: "#/components/schemas/DatabaseConfig"
        errorHandling:
          $ref: "#/components/schemas/ErrorHandlingConfig"
        checks:
          type: array
          items:
            $ref: "#/components/schemas/GoldenPathRule"
      required: [techStack, folderStructure, security, codeQuality, database, errorHandling, checks]
    GoldenPathConfigResponse:
      type: object
      properties:
        id:
          type: string
          nullable: true
        name:
          type: string
        description:
          type: string
          nullable: true
        rules:
          $ref: "#/components/schemas/GoldenPathConfigRules"
        isActive:
          type: boolean
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [name, rules, isActive, isDefault]
    GoldenPathConfigListResponse:
      type: object
      properties:
        configs:
          type: array
          items:
            $ref: "#/components/schemas/GoldenPathConfigResponse"
        defaultRules:
          $ref: "#/components/schemas/GoldenPathConfigRules"
      required: [configs, defaultRules]
    CreateGoldenPathConfigBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        rules:
          $ref: "#/components/schemas/GoldenPathConfigRules"
        isActive:
          type: boolean
      required: [name, rules]
    UpdateGoldenPathConfigBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        rules:
          $ref: "#/components/schemas/GoldenPathConfigRules"
        isActive:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error